refactor(ref_select): Move RefLinkEnsurer into ref_select.go
