+++ cmd/ref_select.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/qri-io/qri/dsref"
 	"github.com/qri-io/qri/fsi"
+	"github.com/qri-io/qri/lib"
 	"github.com/qri-io/qri/repo"
 )
 
@@ -98,12 +99,14 @@ func (r *RefSelect) String() string {
 	return fmt.Sprintf("%s dataset [%s]", r.kind, strings.Join(r.refs, ", "))
 }
 
-// AnyNumberOfReferences is for commands that can work on any number of dataset references
-const AnyNumberOfReferences = -1
+const (
+	// AnyNumberOfReferences is for commands that can work on any number of dataset references
+	AnyNumberOfReferences = -1
 
-// BadUpperCaseOkayWhenSavingExistingDataset is for the save command, which can have bad
-// upper-case characters in its reference but only if it already exists
-const BadUpperCaseOkayWhenSavingExistingDataset = -2
+	// BadUpperCaseOkayWhenSavingExistingDataset is for the save command, which can have bad
+	// upper-case characters in its reference but only if it already exists
+	BadUpperCaseOkayWhenSavingExistingDataset = -2
+)
 
 // GetCurrentRefSelect returns the current reference selection. This could be explicitly provided
 // as command-line arguments, or could be determined by being in a linked directory, or could be
@@ -198,4 +201,40 @@ func DefaultSelectedRefList(f Factory) ([]string, error) {
 	}
 
 	return res, nil
+}
+
+// EnsureFSIAgrees should be passed to GetCurrentRefSelect in order to ensure that any references
+// used by a command have agreement between what their .qri-ref linkfile thinks and what the
+// qri repository thinks. If there's a disagreement, the linkfile wins and the repository will
+// be updated to match.
+// This is useful if a user has a working directory, and then manually deletes the .qri-ref (which
+// will unlink the dataset), or renames / moves the directory and then runs a command in that
+// directory (which will update the repository with the new working directory's path).
+func EnsureFSIAgrees(f *lib.FSIMethods) *FSIRefLinkEnsurer {
+	if f == nil {
+		return nil
+	}
+	return &FSIRefLinkEnsurer{FSIMethods: f}
+}
+
+// FSIRefLinkEnsurer is a simple wrapper for ensuring the linkfile agrees with the repository. We
+// use it instead of a raw FSIMethods pointer so that users of this code see they need to call
+// EnsureFSIAgrees(*fsiMethods) when calling GetRefSelect, hopefully providing a bit of insight
+// about what this parameter is for.
+type FSIRefLinkEnsurer struct {
+	FSIMethods *lib.FSIMethods
+}
+
+// EnsureRef checks if the linkfile and repository agree on the dataset's working directory path.
+// If not, it will modify the repository so that it matches the linkfile. The linkfile will
+// never be modified.
+func (e *FSIRefLinkEnsurer) EnsureRef(refs *RefSelect) error {
+	if e == nil {
+		return nil
+	}
+	p := lib.EnsureParams{Dir: refs.Dir(), Ref: refs.Ref()}
+	info := dsref.VersionInfo{}
+	// Lib call matches the gorpc method signature, but `out` is not used
+	err := e.FSIMethods.EnsureRef(&p, &info)
+	return err
 }
